requirements.txt
flask==2.3.3
sqlalchemy==2.0.23
llama-cpp-python==0.2.19
Werkzeug==2.3.7
gunicorn==21.2.0 